# api-contract.yaml
openapi: 3.0.3
info:
  title: 小猫月度预算账本 API
  version: 1.1.0-draft
  description: |
    小猫月度预算账本服务端接口规范。

    ## 认证
    小程序接口使用 Bearer JWT；Web Agent 使用 HttpOnly Cookie，并为非安全请求提交 `X-CSRF-TOKEN`。
    `Authorization: Bearer <token>`

    ## 统一响应体
    所有接口返回 `Result<T>`，分页接口返回 `PageResult<T>`：
    - 成功：`{ "code": 200, "message": "success", "data": ... }`
    - 失败：`{ "code": 4xx/5xx, "message": "错误描述", "data": null }`

    ## 金额说明
    金额统一使用字符串传输，后端使用 `DECIMAL(12,2)` 存储，精确到分。

    ## 日期格式
    - 年月：`yyyy-MM`（如 `2026-08`）
    - 日期：`yyyy-MM-dd`（如 `2026-08-23`）

servers:
  - url: /api
    description: 服务端 API 根路径

security:
  - bearerAuth: []

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: WEB_AUTH

  schemas:
    # ─── 通用响应 ────────────────────────────────────────────────
    Result:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: success
        data:
          description: 业务数据，失败时为 null

    PageResult:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: success
        data:
          type: object
          required: [total, page, pageSize, list]
          properties:
            total:
              type: integer
              description: 总记录数
            page:
              type: integer
              description: 当前页（从 1 开始）
            pageSize:
              type: integer
              description: 每页条数
            list:
              type: array
              items: {}

    # ─── 错误码约定 ───────────────────────────────────────────────
    # 200  成功
    # 400  参数错误（BizException 业务异常）
    # 401  未认证（JWT 缺失或过期）
    # 403  无权限
    # 404  资源不存在
    # 500  服务内部错误

    # ─── 用户 ─────────────────────────────────────────────────────
    LoginRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
          description: 微信 wx.login() 返回的 code

    AvatarKey:
      type: string
      description: 应用内置卡通小猫头像标识。客户端根据该标识展示本地头像资源。
      enum: [cat-orange, cat-black, cat-calico, cat-gray, cat-white, cat-ragdoll, cat-tuxedo, cat-siamese, cat-tabby]
      default: cat-orange

    LoginResponse:
      type: object
      required: [token, userId, nickname, avatar]
      properties:
        token:
          type: string
          description: JWT，有效期 7 天
        userId:
          type: integer
          format: int64
        nickname:
          type: string
        avatar:
          $ref: '#/components/schemas/AvatarKey'

    UserProfile:
      type: object
      required: [userId, nickname, avatar, createdAt]
      properties:
        userId:
          type: integer
          format: int64
        nickname:
          type: string
        avatar:
          $ref: '#/components/schemas/AvatarKey'
        phone:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time

    UpdateProfileRequest:
      type: object
      properties:
        nickname:
          type: string
          maxLength: 20
        avatar:
          $ref: '#/components/schemas/AvatarKey'

    # ─── 科目（Category） ─────────────────────────────────────────
    Category:
      type: object
      required: [categoryId, name, isSystem]
      properties:
        categoryId:
          type: integer
          format: int64
        name:
          type: string
          example: 饮食
        isSystem:
          type: boolean
          description: true 表示系统预置科目，不可删除

    CreateCategoryRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 20
          description: 科目名称，同一用户不允许重名（含系统预置名称）；名称已存在时直接返回已有科目，不重复创建

    # ─── 月度预算计划（Budget Plan） ──────────────────────────────
    BudgetItem:
      type: object
      required: [categoryId, amount, sortOrder]
      properties:
        categoryId:
          type: integer
          format: int64
        categoryName:
          type: string
        amount:
          type: string
          description: 预算金额，正数，保留两位小数
          pattern: '^\d{1,10}(\.\d{1,2})?$'
          example: "2000.00"
        sortOrder:
          type: integer
          description: 科目排序序号，按月独立存储，从 0 开始
          example: 0

    SaveBudgetPlanRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          description: |
            本月所有科目的预算明细。调用此接口会覆盖该月已有预算（先删除再插入，事务保证）。
            同一科目不允许重复出现。金额必须 > 0。
          items:
            $ref: '#/components/schemas/BudgetItem'

    BudgetPlan:
      type: object
      required: [yearMonth, items, totalBudget]
      properties:
        yearMonth:
          type: string
          example: "2026-08"
        totalBudget:
          type: string
          description: 各科目预算之和
          example: "8000.00"
        items:
          type: array
          items:
            $ref: '#/components/schemas/BudgetItem'

    BudgetSummaryItem:
      type: object
      required: [categoryId, categoryName, budget, actual, remaining, rate]
      properties:
        categoryId:
          type: integer
          format: int64
        categoryName:
          type: string
        budget:
          type: string
          description: 预算金额
          example: "2000.00"
        actual:
          type: string
          description: 实际已花费
          example: "2450.00"
        remaining:
          type: string
          description: 剩余额度（可为负数，表示超支）
          example: "-450.00"
        rate:
          type: number
          format: float
          description: 执行率（actual / budget），百分比值
          example: 122.5

    BudgetSummary:
      type: object
      required: [yearMonth, totalBudget, totalActual, totalRemaining, rate, items]
      properties:
        yearMonth:
          type: string
          example: "2026-08"
        totalBudget:
          type: string
          example: "8000.00"
        totalActual:
          type: string
          example: "5820.00"
        totalRemaining:
          type: string
          example: "2180.00"
        rate:
          type: number
          format: float
          description: 总执行率（百分比）
          example: 72.75
        items:
          type: array
          items:
            $ref: '#/components/schemas/BudgetSummaryItem'

    # ─── 开销记录（Expense） ──────────────────────────────────────
    Expense:
      type: object
      required: [expenseId, categoryId, categoryName, amount, expenseDate, createdAt]
      properties:
        expenseId:
          type: integer
          format: int64
        categoryId:
          type: integer
          format: int64
        categoryName:
          type: string
          description: 快照字段，即使科目已删除也返回原始名称
        amount:
          type: string
          description: 金额，保留两位小数
          example: "68.50"
        expenseDate:
          type: string
          format: date
          example: "2026-08-23"
        note:
          type: string
          nullable: true
          maxLength: 100
        createdAt:
          type: string
          format: date-time

    CreateExpenseRequest:
      type: object
      required: [categoryId, amount, expenseDate]
      properties:
        categoryId:
          type: integer
          format: int64
          description: 必须是可用状态的科目（系统预置或用户已创建的科目）
        amount:
          type: string
          description: 金额，必须 > 0，保留两位小数
          pattern: '^\d{1,10}(\.\d{1,2})?$'
          example: "68.50"
        expenseDate:
          type: string
          format: date
          description: 消费日期，格式 yyyy-MM-dd，不允许晚于今天
          example: "2026-08-23"
        note:
          type: string
          maxLength: 100
          nullable: true

    UpdateExpenseRequest:
      type: object
      properties:
        categoryId:
          type: integer
          format: int64
          description: 修改科目，必须是可用状态的科目（系统预置或用户已创建的科目）
        amount:
          type: string
          description: 金额，必须 > 0，保留两位小数
          pattern: '^\d{1,10}(\.\d{1,2})?$'
          example: "68.50"
        expenseDate:
          type: string
          format: date
          description: 修改消费日期，格式 yyyy-MM-dd，不允许晚于今天
          example: "2026-08-23"
        note:
          type: string
          maxLength: 100
          nullable: true

    ExpenseStatItem:
      type: object
      required: [categoryId, categoryName, total, count, percentage]
      properties:
        categoryId:
          type: integer
          format: int64
        categoryName:
          type: string
        total:
          type: string
          description: 该科目总金额
          example: "2450.00"
        count:
          type: integer
          description: 记录条数
        percentage:
          type: number
          format: float
          description: 占当月总开销的百分比
          example: 42.1

    ExpenseStats:
      type: object
      required: [yearMonth, grandTotal, totalCount, dailyAvg, items]
      properties:
        yearMonth:
          type: string
          example: "2026-08"
        grandTotal:
          type: string
          description: 当月总开销
          example: "5820.00"
        totalCount:
          type: integer
          description: 当月总笔数
          example: 42
        dailyAvg:
          type: string
          description: 日均支出（grandTotal / 当月自然天数）
          example: "187.74"
        items:
          type: array
          items:
            $ref: '#/components/schemas/ExpenseStatItem'

    TrendItem:
      type: object
      required: [yearMonth, total]
      properties:
        yearMonth:
          type: string
          example: "2026-08"
        total:
          type: string
          example: "5820.00"

# ─── 路径定义 ──────────────────────────────────────────────────────

    WebLoginRequest:
      type: object
      required: [requestId, browserProof, mode, expiresAt]
      properties:
        requestId: { type: string }
        browserProof: { type: string, description: 仅浏览器持有的交换凭据 }
        mode: { type: string, enum: [DYNAMIC_QR, FALLBACK_CODE] }
        fallbackCode: { type: string, pattern: '^\\d{6}$', nullable: true }
        fixedQrCodeUrl: { type: string, format: uri, nullable: true, description: 固定小程序码图片地址，回退模式使用 }
        qrCodeUrl: { type: string, format: uri, nullable: true }
        expiresAt: { type: string, format: date-time }
      oneOf:
        - required: [qrCodeUrl]
          properties: { mode: { enum: [DYNAMIC_QR] }, fallbackCode: { nullable: true } }
        - required: [fallbackCode, fixedQrCodeUrl]
          properties: { mode: { enum: [FALLBACK_CODE] }, qrCodeUrl: { nullable: true } }
    WebLoginStatus:
      type: object
      properties:
        requestId: { type: string }
        deviceLabel: { type: string }
        status: { type: string, enum: [PENDING, APPROVED, REJECTED, EXPIRED, CONSUMED] }
        expiresAt: { type: string, format: date-time }
    FallbackCodeRequest:
      type: object
      required: [fallbackCode]
      properties:
        fallbackCode: { type: string, pattern: '^\d{6}$' }
    BrowserProofRequest:
      type: object
      required: [browserProof]
      properties:
        browserProof: { type: string, minLength: 1, description: 仅创建请求的浏览器持有 }
    SsoExchangeRequest:
      type: object
      required: [ticket]
      properties:
        ticket: { type: string, minLength: 1, description: URL fragment 中的 60 秒一次性票据 }
    SsoLink:
      type: object
      required: [loginUrl, expiresAt]
      properties:
        loginUrl: { type: string, format: uri }
        expiresAt: { type: string, format: date-time }
    AgentChatRequest:
      type: object
      required: [message]
      properties:
        sessionId: { type: string, format: uuid, nullable: true }
        message: { type: string, minLength: 1, maxLength: 2000 }
        thinkingEnabled: { type: boolean, default: false }
        retryUserMessageId: { type: string, format: uuid, nullable: true, description: 失败重试时复用的原 USER 消息；必须属于同一用户与会话且正文一致 }
    AgentSession:
      type: object
      properties:
        sessionId: { type: string, format: uuid }
        title: { type: string }
        updatedAt: { type: string, format: date-time }
    AgentMessage:
      type: object
      properties:
        messageId: { type: string, format: uuid }
        role: { type: string, enum: [USER, ASSISTANT] }
        content: { type: string }
        action: { $ref: '#/components/schemas/AgentAction', nullable: true }
        status: { type: string, enum: [COMPLETED, FAILED] }
        createdAt: { type: string, format: date-time }
    AgentAction:
      type: object
      required: [actionId, type, summary, payload, status, expiresAt]
      properties:
        actionId: { type: string, format: uuid }
        type: { type: string, enum: [CREATE_EXPENSE, UPDATE_EXPENSE, DELETE_EXPENSE, SAVE_PLAN, CREATE_CATEGORY] }
        summary: { type: string }
        payload: { type: object, additionalProperties: true }
        result: { type: object, additionalProperties: true, nullable: true, description: 执行结果或可读失败原因 }
        status: { type: string, enum: [PENDING_CONFIRMATION, EXECUTING, CANCELLED, EXPIRED, STALE, EXECUTED, FAILED] }
        expiresAt: { type: string, format: date-time }

paths:
  /web-auth/requests:
    post:
      security: []
      summary: 创建电脑端网页登录请求（动态码或固定码兜底）
      responses:
        '200': { description: 成功；data 为 WebLoginRequest，动态码与固定码字段互斥 }
        '429': { description: 创建过于频繁 }
        '503': { description: 固定码配置缺失或暂时无法分配登录码 }
  /web-auth/requests/{requestId}/status:
    get:
      security: []
      parameters: [{ name: requestId, in: path, required: true, schema: { type: string } }, { name: X-Web-Login-Proof, in: header, required: true, schema: { type: string } }]
      responses:
        '200': { description: data 为 WebLoginStatus }
        '404': { description: 请求不存在或浏览器凭据错误 }
  /web-auth/requests/{requestId}/qr:
    get:
      security: []
      summary: 获取动态小程序码
      parameters: [{ name: requestId, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: PNG 图片, content: { image/png: { schema: { type: string, format: binary } } } } }
  /web-auth/requests/{requestId}/preview:
    get:
      summary: 小程序预览登录请求
      parameters: [{ name: requestId, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: data 为 WebLoginStatus }, '404': { description: 请求不存在或已不可确认 } }
  /web-auth/requests/{requestId}/approve:
    post:
      parameters: [{ name: requestId, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: 已确认 } }
  /web-auth/requests/{requestId}/reject:
    post:
      parameters: [{ name: requestId, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: 已拒绝 } }
  /web-auth/fallback/resolve:
    post:
      summary: 小程序通过六位码解析登录请求
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/FallbackCodeRequest' } } } }
      responses: { '200': { description: data 为 WebLoginStatus }, '404': { description: 登录码不存在或已失效 }, '429': { description: 尝试过于频繁 } }
  /web-auth/miniapp-links:
    post:
      summary: 小程序生成一次性 Web 登录链接
      responses: { '200': { description: data 为 SsoLink }, '401': { description: 小程序身份无效 } }
  /web-auth/sso/exchange:
    post:
      security: []
      summary: 一次性短票据交换 Web Cookie
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/SsoExchangeRequest' } } } }
      responses: { '200': { description: 设置 HttpOnly WEB_AUTH 和可读 XSRF-TOKEN Cookie }, '409': { description: 票据过期或已使用 } }
  /web-auth/requests/{requestId}/exchange:
    post:
      security: []
      parameters: [{ name: requestId, in: path, required: true, schema: { type: string } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/BrowserProofRequest' } } } }
      responses: { '200': { description: 设置 HttpOnly WEB_AUTH 和可读 XSRF-TOKEN Cookie }, '401': { description: 浏览器凭据错误 }, '409': { description: 请求未确认、过期、拒绝或已使用 } }
  /agent/chat:
    post:
      security: [{ cookieAuth: [] }]
      summary: 发送 Agent 消息并接收 SSE
      description: 仅发送 message_start、delta、action_required、done、error；不发送推理内容或内部工具事件。429、总超时和断流以 error 事件或对应 HTTP 状态表示，均可重试。
      parameters: [{ name: X-CSRF-TOKEN, in: header, required: true, schema: { type: string } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AgentChatRequest' } } } }
      responses: { '200': { description: text/event-stream }, '403': { description: CSRF Token 缺失或错误 }, '429': { description: 用户分钟限流 }, '503': { description: DeepSeek 不可用 } }
  /agent/sessions:
    get:
      security: [{ cookieAuth: [] }]
      responses: { '200': { description: 'Result<AgentSession[]>' } }
  /agent/sessions/{sessionId}/messages:
    get:
      security: [{ cookieAuth: [] }]
      parameters: [{ name: sessionId, in: path, required: true, schema: { type: string, format: uuid } }]
      responses: { '200': { description: 'Result<AgentMessage[]>，仅 USER/ASSISTANT 可见消息' }, '404': { description: 会话不存在或不属于当前用户 } }
  /agent/actions/{actionId}/confirm:
    post:
      security: [{ cookieAuth: [] }]
      parameters: [{ name: actionId, in: path, required: true, schema: { type: string, format: uuid } }, { name: X-CSRF-TOKEN, in: header, required: true, schema: { type: string } }]
      responses: { '200': { description: 'Result<AgentAction>；返回 EXECUTED、STALE、EXPIRED、FAILED 或原状态' }, '403': { description: CSRF Token 缺失或错误 }, '404': { description: 动作不存在或不属于当前用户 } }
  /agent/actions/{actionId}/cancel:
    post:
      security: [{ cookieAuth: [] }]
      parameters: [{ name: actionId, in: path, required: true, schema: { type: string, format: uuid } }, { name: X-CSRF-TOKEN, in: header, required: true, schema: { type: string } }]
      responses: { '200': { description: 返回取消后的动作状态 } }

  # ── 用户 ──────────────────────────────────────────────────────────
  /user/login:
    post:
      tags: [用户]
      summary: 微信登录（获取 JWT）
      description: |
        传入微信 wx.login() 返回的 code，服务端调用 code2session 换取 openid，
        若用户不存在则自动注册，返回 JWT。
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Result'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LoginResponse'
        '400':
          description: code 无效或 code2session 调用失败

  /user/profile:
    get:
      tags: [用户]
      summary: 获取当前用户信息
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Result'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserProfile'
        '401':
          description: JWT 无效、过期，或令牌对应用户已不存在

    put:
      tags: [用户]
      summary: 更新用户昵称/内置头像
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: 更新成功，返回更新后的用户信息
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Result'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserProfile'
        '401':
          description: JWT 无效、过期，或令牌对应用户已不存在

  # ── 科目管理 ───────────────────────────────────────────────────────
  /categories:
    get:
      tags: [科目]
      summary: 获取可用科目列表
      description: 返回当前用户可用的全部科目：系统预置科目 + 用户自定义科目。
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Result'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Category'

    post:
      tags: [科目]
      summary: 新建自定义科目
      description: |
        科目名称在同一用户下不允许重名（含系统预置名称）。
        名称已存在时直接返回已有科目对象（HTTP 200），不重复创建。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryRequest'
      responses:
        '200':
          description: 创建成功或已存在，返回科目对象
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Result'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Category'
        '400':
          description: 参数不合法（名称为空、超长等）

  # ── 月度预算计划 ────────────────────────────────────────────────────
  /plan/has-any:
    get:
      tags: [预算计划]
      summary: 检查是否设置过任何月份的预算
      description: |
        检查当前用户是否曾设置过任意月份的预算，用于记账入口拦截。
        从未设置过预算时返回 false，强制跳转预算编辑页。
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Result'
                  - type: object
                    properties:
                      data:
                        type: boolean
                        description: true 表示曾设置过预算；false 表示从未设置过

  /plan/{yearMonth}:
    parameters:
      - name: yearMonth
        in: path
        required: true
        schema:
          type: string
          pattern: '^\d{4}-(0[1-9]|1[0-2])$'
        example: "2026-08"

    post:
      tags: [预算计划]
      summary: 创建或覆盖月度预算
      description: |
        保存指定月份的预算计划。使用覆盖语义：
        先在事务中删除该月旧数据，再插入新数据。
        同一科目在 items 数组中不允许重复出现。
        yearMonth 不得为未来月份，否则返回 400。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveBudgetPlanRequest'
      responses:
        '200':
          description: 保存成功，返回保存后的预算计划
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Result'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BudgetPlan'
        '400':
          description: 参数不合法（金额<=0、科目重复、科目不可用等）

    get:
      tags: [预算计划]
      summary: 查询月度预算详情
      description: 若该月尚无预算，data 为 null；客户端可据此查找最近的历史预算用于继承。
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Result'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BudgetPlan'
                        nullable: true

  /plan/{yearMonth}/summary:
    get:
      tags: [预算计划]
      summary: 预算 vs 实际开销汇总
      description: |
        返回指定月份各科目的预算执行情况：预算金额、实际花费、剩余、执行率。
        若该月无预算记录，仍返回实际开销；totalBudget 为 "0.00"，remaining 为负实际开销。
        有开销但无预算的科目，budget 字段返回 "0.00"，remaining 为负值。
      parameters:
        - name: yearMonth
          in: path
          required: true
          schema:
            type: string
            pattern: '^\d{4}-(0[1-9]|1[0-2])$'
          example: "2026-08"
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Result'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BudgetSummary'
                        nullable: true

  # ── 开销记录 ─────────────────────────────────────────────────────────
  /expense:
    get:
      tags: [开销记录]
      summary: 分页查询开销列表
      parameters:
        - name: yearMonth
          in: query
          required: true
          schema:
            type: string
            pattern: '^\d{4}-(0[1-9]|1[0-2])$'
          example: "2026-08"
          description: 按年月筛选（必传）
        - name: categoryId
          in: query
          required: false
          schema:
            type: integer
            format: int64
          description: 按科目筛选（可选）
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: 成功，按 expenseDate 倒序排列
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageResult'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          list:
                            type: array
                            items:
                              $ref: '#/components/schemas/Expense'

    post:
      tags: [开销记录]
      summary: 新增开销记录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExpenseRequest'
      responses:
        '200':
          description: 创建成功，返回新记录
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Result'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Expense'
        '400':
          description: 参数不合法（金额<=0、科目已删除等）

  /expense/{expenseId}:
    parameters:
      - name: expenseId
        in: path
        required: true
        schema:
          type: integer
          format: int64

    put:
      tags: [开销记录]
      summary: 编辑开销记录
      description: |
        只更新传入的字段（PATCH 语义，但使用 PUT 方法）。
        若传入 categoryId，该科目必须是可用状态。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateExpenseRequest'
      responses:
        '200':
          description: 更新成功，返回更新后的记录
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Result'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Expense'
        '400':
          description: 参数不合法
        '404':
          description: 记录不存在或不属于当前用户

    delete:
      tags: [开销记录]
      summary: 软删除开销记录
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        '404':
          description: 记录不存在或不属于当前用户

  /expense/stats/{yearMonth}:
    get:
      tags: [开销记录]
      summary: 月度按科目汇总统计
      description: 返回指定月份各科目的开销总额、笔数和占比，按 total 降序排列。
      parameters:
        - name: yearMonth
          in: path
          required: true
          schema:
            type: string
            pattern: '^\d{4}-(0[1-9]|1[0-2])$'
          example: "2026-08"
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Result'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ExpenseStats'

  /expense/trend:
    get:
      tags: [开销记录]
      summary: 近 N 月开销趋势
      parameters:
        - name: months
          in: query
          required: false
          schema:
            type: integer
            default: 6
            minimum: 1
            maximum: 24
          description: 返回最近 N 个月的数据，含当月
      responses:
        '200':
          description: 成功，按月份升序排列
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Result'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/TrendItem'